[EI-227] [web] feat: Add vertical split screen component
